Started refactor+tidy firecracker runtime plugin
