--- pkg/snapshotter/errors.go
@@ -1,30 +1,0 @@
-package snapshotter
-
-import "errors"
-
-var (
-	ErrCouldNotGetDeviceStat                 = errors.New("could not get NBD device stat")
-	ErrCouldNotWaitForFirecrackerServer      = errors.New("could not wait for Firecracker server")
-	ErrCouldNotOpenLivenessServer            = errors.New("could not open liveness server")
-	ErrCouldNotChownLivenessServerVSock      = errors.New("could not change ownership of liveness server VSock")
-	ErrCouldNotChownAgentServerVSock         = errors.New("could not change ownership of agent server VSock")
-	ErrCouldNotOpenInputFile                 = errors.New("could not open input file")
-	ErrCouldNotCreateOutputFile              = errors.New("could not create output file")
-	ErrCouldNotCopyFile                      = errors.New("error copying file")
-	ErrCouldNotWritePadding                  = errors.New("could not write padding")
-	ErrCouldNotCopyDeviceFile                = errors.New("could not copy device file")
-	ErrCouldNotStartVM                       = errors.New("could not start VM")
-	ErrCouldNotReceiveAndCloseLivenessServer = errors.New("could not receive and close liveness server")
-	ErrCouldNotAcceptAgentConnection         = errors.New("could not accept agent connection")
-	ErrCouldNotBeforeSuspend                 = errors.New("error before suspend")
-	ErrCouldNotMarshalPackageConfig          = errors.New("could not marshal package configuration")
-	ErrCouldNotOpenPackageConfigFile         = errors.New("could not open package configuration file")
-	ErrCouldNotWritePackageConfig            = errors.New("could not write package configuration")
-	ErrCouldNotChownPackageConfigFile        = errors.New("could not change ownership of package configuration file")
-	ErrCouldNotCreateChrootBaseDirectory     = errors.New("could not create chroot base directory")
-	ErrCouldNotStartFirecrackerServer        = errors.New("could not start firecracker server")
-	ErrCouldNotStartAgentServer              = errors.New("could not start agent server")
-	ErrCouldNotWaitForAcceptingAgent         = errors.New("could not wait for accepting agent")
-	ErrCouldNotCloseAcceptingAgent           = errors.New("could not close accepting agent")
-	ErrCouldNotCreateSnapshot                = errors.New("could not create snapshot")
-)
